backend/requirements.txt
fastapi
uvicorn
pymilvus
python-multipart
sqlalchemy
aiosqlite
openai
pydantic-settings
langchain-text-splitters
pypdf
tiktoken
sentence-transformers
greenlet
langchain-experimental
langchain-openai
rank-bm25
numpy
spacy
rdflib>=7.0.0
SPARQLWrapper>=2.0.0
httpx>=0.24.0
